adding the factory for reading as well
