+++ controllers/handlerFactory.js
@@ -1,4 +1,5 @@
 const catchAsync = require('./../utils/catchAsync');
+const APIFeatures = require('./../utils/apiFeatures');
 const AppError = require('./../utils/appError');
 
 exports.deleteOne = Model =>
@@ -40,6 +41,49 @@ exports.createOne = Model =>
 
     res.status(201).json({
       status: 'success',
+      data: {
+        data: doc
+      }
+    });
+  });
+
+exports.getOne = (Model, popOptions) =>
+  catchAsync(async (req, res, next) => {
+    let query = Model.findById(req.params.id);
+    if (popOptions) query = query.populate(popOptions);
+    const doc = await query;
+
+    if (!doc) {
+      return next(new AppError('No document found with that ID!', 404));
+    }
+
+    res.status(200).json({
+      status: 'success',
+      data: {
+        data: doc
+      }
+    });
+  });
+
+exports.getAll = Model =>
+  catchAsync(async (req, res, next) => {
+    // To allow for nested GET reviews on course
+    let filter = {};
+    if (req.params.courseId) filter = { course: req.params.courseId };
+
+    // EXECUTE QUERY
+    const features = new APIFeatures(Model.find(filter), req.query)
+      .filter()
+      .sort()
+      .limitFields()
+      .paginate();
+    const doc = await features.query;
+    // query.sort().select().skip().limit()
+
+    // SEND RESPONSE
+    res.status(200).json({
+      status: 'success',
+      results: doc.length,
       data: {
         data: doc
       }
