+++ controllers/reviewController.js
@@ -1,21 +1,21 @@
 const Review = require('./../models/reviewModel');
-const catchAsync = require('./../utils/catchAsync');
+// const catchAsync = require('./../utils/catchAsync');
 const factory = require('./handlerFactory');
 
-exports.getAllReviews = catchAsync(async (req, res, next) => {
-  let filter = {};
-  if (req.params.courseId) filter = { course: req.params.courseId };
+// exports.getAllReviews = catchAsync(async (req, res, next) => {
+//   let filter = {};
+//   if (req.params.courseId) filter = { course: req.params.courseId };
 
-  const reviews = await Review.find(filter);
+//   const reviews = await Review.find(filter);
 
-  res.status(200).json({
-    status: 'success',
-    results: reviews.length,
-    data: {
-      reviews
-    }
-  });
-});
+//   res.status(200).json({
+//     status: 'success',
+//     results: reviews.length,
+//     data: {
+//       reviews
+//     }
+//   });
+// });
 
 // exports.createReview = catchAsync(async (req, res, next) => {
 //   // Allow nested routes
@@ -32,13 +32,17 @@ exports.getAllReviews = catchAsync(async (req, res, next) => {
 //   });
 // });
 
-exports.createReview = factory.createOne(Review);
-
 exports.setCourseUserIds = (req, res, next) => {
   if (!req.body.course) req.body.course = req.params.courseId;
   if (!req.body.user) req.body.user = req.user.id;
   next();
 };
+
+exports.getAllReviews = factory.getAll(Review);
+
+exports.getReview = factory.getOne(Review);
+
+exports.createReview = factory.createOne(Review);
 
 exports.updateReview = factory.updateOne(Review);
