+++ routes/reviewRoutes.js
@@ -16,6 +16,7 @@ router
 
 router
   .route('/:id')
+  .get(reviewController.getReview)
   .patch(reviewController.updateReview)
   .delete(reviewController.deleteReview);
